move acf to autocov
